[UHYU-216] refactor: 어드민 관련 메서드 폴더 이동
