+++ src/main/java/com/ureca/uhyu/domain/admin/service/AdminService.java
@@ -1,0 +1,12 @@
+package com.ureca.uhyu.domain.admin.service;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+public class AdminService {
+
+}
